+++ pytorch_ipynb/helper_evaluate.py
@@ -1,0 +1,34 @@
+import torch
+import torch.nn.functional as F
+
+
+def compute_accuracy(model, data_loader, device):
+    model.eval()
+    with torch.no_grad():
+        correct_pred, num_examples = 0, 0
+        for i, (features, targets) in enumerate(data_loader):
+
+            features = features.to(device)
+            targets = targets.to(device)
+
+            logits = model(features)
+            _, predicted_labels = torch.max(logits, 1)
+            num_examples += targets.size(0)
+            correct_pred += (predicted_labels == targets).sum()
+    return correct_pred.float()/num_examples * 100
+
+
+def compute_epoch_loss(model, data_loader, device):
+    model.eval()
+    curr_loss, num_examples = 0., 0
+    with torch.no_grad():
+        for features, targets in data_loader:
+            features = features.to(device)
+            targets = targets.to(device)
+            logits = model(features)
+            loss = F.cross_entropy(logits, targets, reduction='sum')
+            num_examples += targets.size(0)
+            curr_loss += loss
+
+        curr_loss = curr_loss / num_examples
+        return curr_loss
